backend/requirements.txt
# FastAPI + Server
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
sse-starlette==2.2.1

# Database (PostgreSQL + pgvector)
sqlalchemy==2.0.41
alembic==1.15.2
asyncpg==0.30.0
pgvector==0.3.6

# Auth
pyjwt==2.10.1
bcrypt==4.3.0

# Settings
pydantic-settings==2.9.1
email-validator==2.3.0

# AI / RAG
google-genai==1.14.0
openai==1.82.0
cohere==5.13.0

# Document Processing
pymupdf==1.27.2
python-docx==1.1.2
beautifulsoup4==4.13.4

# HTTP
httpx==0.28.1

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
